requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
pydantic
python-dotenv
python-multipart
pypdf
streamlit
requests